Corrected HP issues in PlayerSwitch
PlayerSwitch was inconsistent with regards to the HP of the player and cloned peds. GTA5 by default refills the player's HP when switching, but this is a bug, not a feature. Updated PlayerSwitch will now correctly set the player and cloned ped HP levels on switch. If the old functionality is desired, an optional bool fullHealth is provided for both Switch and CreatePedAndSwitch to allow switched peds to regain full HP as before.