Wanted level change to MomentReplica
Added bool to allow a MomentReplica to determine whether the player's wanted level should be changed or not. This is handy for using MomentReplica as a weather event which should not affect a player's wanted level when applied. Defaults to true for backwards-compatibility with functions that expect the original functionality.