Add NoMods SpawnFlag for VehicleReplica
Adds a SpawnFlag to prevent VehicleReplica from interacting with a vehicle's Mods or Extras list. Used for situations where the target vehicle already has Extras and Mods that are under direct control of a running script and as such VehicleReplica attempting to apply them would cause a conflict as both scripts attempt to change the same parameters every single frame.