Adjust Ground-related Functions
Removes hardcoded offset in GetPositionOnGround, replaces hardcoded raycast in IsWheelOnTracks with one that accounts for variability in wheel height, and adds a function called IsOutInTheOpen that checks to see if a given entity is out in the open, i.e. not obscured by anything overhead. If the given entity is either underground (for example, in a metro station) or has some sort of structure overhead (such as in a parking garage), this will return false. Otherwise, it returns true. Useful for making sure peds are above ground, and if vehicles are in open areas.