Update ModelSwap
Updated ModelSwap code to make it a little closer to reality. Start of production is not start of commercial availability, so added 4 months to that check to account for time delay between production and arrival at dealerships. End of production plus 5 years is not the last time you will ever see a car on the road, so changed it so that after 10 years beyond end of production, the chance multiplier will be locked at 0.2f instead of 0. Also switched the seconds calculation to days as calculating seconds for 5/10 years was causing microstutters while driving rapidly through ambient traffic. Also fixed a bug in the Remap function that was making it less likely to see a car the closer to end of production it was rather than more likely.